Extract module `alloc.py` (#721)
