+++ kmir/src/kmir/smir.py
@@ -2,24 +2,25 @@ from __future__ import annotations
 
 import json
 import logging
-from abc import ABC
 from collections import deque
-from dataclasses import dataclass
 from functools import cached_property
-from typing import TYPE_CHECKING, NamedTuple, NewType
+from typing import TYPE_CHECKING, NewType
 
+from .alloc import AllocInfo
 from .ty import EnumT, RefT, StructT, Ty, TypeMetadata, UnionT
 
 if TYPE_CHECKING:
     from pathlib import Path
-    from typing import Any, Final
+    from typing import Final
 
+    from .alloc import AllocId
 
-AllocId = NewType('AllocId', int)
-AdtDef = NewType('AdtDef', int)
 
 _LOGGER: Final = logging.getLogger(__name__)
 _LOG_FORMAT: Final = '%(levelname)s %(asctime)s %(name)s - %(message)s'
+
+
+AdtDef = NewType('AdtDef', int)
 
 
 class SMIRInfo:
@@ -219,81 +220,3 @@ def compute_closure(start: Ty, edges: dict[Ty, set[Ty]]) -> set[Ty]:
             if next in edges:
                 work.extend(edges[next])
     return reached
-
-
-@dataclass
-class AllocInfo:
-    alloc_id: AllocId
-    ty: Ty
-    global_alloc: GlobalAlloc
-
-    @staticmethod
-    def from_dict(dct: dict[str, Any]) -> AllocInfo:
-        return AllocInfo(
-            alloc_id=AllocId(dct['alloc_id']),
-            ty=Ty(dct['ty']),
-            global_alloc=GlobalAlloc.from_dict(dct['global_alloc']),
-        )
-
-
-class GlobalAlloc(ABC):  # noqa: B024
-    @staticmethod
-    def from_dict(dct: dict[str, Any]) -> GlobalAlloc:
-        match dct:
-            case {'Memory': _}:
-                return Memory.from_dict(dct)
-            case _:
-                raise ValueError('Unsupported or invalid GlobalAlloc data: {dct}')
-
-
-@dataclass
-class Memory(GlobalAlloc):
-    allocation: Allocation
-
-    @staticmethod
-    def from_dict(dct: dict[str, Any]) -> Memory:
-        return Memory(
-            allocation=Allocation.from_dict(dct['Memory']),
-        )
-
-
-@dataclass
-class Allocation:
-    bytez: list[int | None]  # field 'bytes'
-    provenance: ProvenanceMap
-    align: int
-    mutable: bool  # field 'mutability'
-
-    @staticmethod
-    def from_dict(dct: dict[str, Any]) -> Allocation:
-        return Allocation(
-            bytez=list(dct['bytes']),
-            provenance=ProvenanceMap.from_dict(dct['provenance']),
-            align=int(dct['align']),
-            mutable={
-                'Not': False,
-                'Mut': True,
-            }[dct['mutability']],
-        )
-
-
-@dataclass
-class ProvenanceMap:
-    ptrs: list[ProvenanceItem]
-
-    @staticmethod
-    def from_dict(dct: dict[str, Any]) -> ProvenanceMap:
-        return ProvenanceMap(
-            ptrs=[
-                ProvenanceItem(
-                    size=int(size),
-                    prov=AllocId(prov),
-                )
-                for size, prov in dct['ptrs']
-            ],
-        )
-
-
-class ProvenanceItem(NamedTuple):
-    size: int
-    prov: AllocId
